Split up script in different classes for easier and cleaner development
